drop down, outer component calls context, useful for testing the inner component
